Remove unused wallet repository implementations, exceptions, and utilities to clean up codebase and reduce unused dependencies

+++ domain/user/src/main/java/com/profpay/domain/user/model/WalletOnboardingData.kt
@@ -32,6 +32,6 @@ data class CentralAddress(
 data class SotAddress(
     val address: String,
     val pubKey: String,
-    val index: Int,
+    val index: Byte,
     val derivationIndex: Int,
 )
