+++ domain/wallet/src/main/java/com/profpay/domain/wallet/model/SotAddressParams.kt
@@ -3,6 +3,6 @@ package com.profpay.domain.wallet.model
 data class SotAddressParams(
     val address: String,
     val pubKey: String,
-    val index: Int,
+    val index: Byte,
     val derivationIndex: Int,
 )
